Add pipeline and fix build/install issues
Some issues had arisen due to changes in Arch Linux packages, and btrfs.
This squashed commit fixes these and adds a new build pipeline.

commit 0924e99c2eb41471b75d985c1e77a3613322d666
Author: Ryan Algar <59636191+ralgar@users.noreply.github.com>
Date:   Tue Dec 13 13:15:52 2022 -0800

    Fix: Remove command disabling swap compression

    Disabling compression is now done implicitly when setting NO_COW on the
    swapfile. Attempting to disable compression after setting NO_COW causes
    an error, causing the whole installation to fail.

commit b73b08d50ab14e428e519dca03be614c33f06652
Author: Ryan Algar <59636191+ralgar@users.noreply.github.com>
Date:   Tue Dec 13 12:46:03 2022 -0800

    Fix: Create out directory as unprivileged user

    This avoids permission denied error when creating sha256sum

commit 7d5ad3643118da795aa46c84e8c862d1a1e2a0c3
Author: Ryan Algar <59636191+ralgar@users.noreply.github.com>
Date:   Tue Dec 13 12:23:05 2022 -0800

    Improve build pipeline

commit 6eb3f7c78d23ae7bed689bdde0a7d7a09a52794b
Author: Ryan Algar <59636191+ralgar@users.noreply.github.com>
Date:   Mon Dec 12 16:35:36 2022 -0800

    Fix: Remove deprecated alacritty opacity option

commit efede049ac211ff03853dd9dc1473e9d61957c37
Author: Ryan Algar <59636191+ralgar@users.noreply.github.com>
Date:   Mon Dec 12 15:13:42 2022 -0800

    Update README

commit d61d1f4b35a7cb4a760f89e66db059a1330beca3
Author: Ryan Algar <59636191+ralgar@users.noreply.github.com>
Date:   Mon Dec 12 15:01:28 2022 -0800

    Add release job to pipeline

commit ec4981e21e6ad000d68e51984afbe7d12a7b2468
Author: Ryan Algar <59636191+ralgar@users.noreply.github.com>
Date:   Mon Dec 12 14:08:33 2022 -0800

    Fix: Remove extra namespace from Docker image path

    Image path was namespace/namespace/project, and is now
    namespace/project.

commit 93964ad8f9fe2fd024659d4ebd670949430f7040
Author: Ryan Algar <59636191+ralgar@users.noreply.github.com>
Date:   Mon Dec 12 14:04:23 2022 -0800

    Fix: Remove deprecated packages

    ipw2100-fw and ipw2200-fw are no longer in the Arch repos. Couldn't find
    much information on this, other than the drivers might have been merged
    into the kernel.

commit 3624792b8c9960b63ee6173ed8a0b18c183a2373
Author: Ryan Algar <59636191+ralgar@users.noreply.github.com>
Date:   Mon Dec 12 13:39:57 2022 -0800

    Add initial build pipeline

commit 4931452dc44a27aae88a5af9597bef5ce30d6b1f
Author: Ryan Algar <59636191+ralgar@users.noreply.github.com>
Date:   Tue Nov 29 15:21:59 2022 -0800

    Update profiledef.sh

commit 5c77c722cce25e550a19f8059308d21e5f42f43d
Author: Ryan Algar <59636191+ralgar@users.noreply.github.com>
Date:   Tue Nov 29 14:03:41 2022 -0800

    Replace build.sh with Makefile

commit def80f6c7599a610f17161f03d01c4e9b8464858
Author: Ryan Algar <59636191+ralgar@users.noreply.github.com>
Date:   Sat Oct 8 17:43:46 2022 -0700

    Fix: Shellcheck errors

commit d201095ba51171e7c99254006e797aab7b957fdb
Author: Ryan Algar <59636191+ralgar@users.noreply.github.com>
Date:   Sat Oct 8 17:18:29 2022 -0700

    Move project source into src/ to keep things tidy

commit e82acd1aa03cf853980d9436c941ca72e653b794
Author: Ryan Algar <59636191+ralgar@users.noreply.github.com>
Date:   Sat Oct 8 17:05:40 2022 -0700

    Fix shellcheck error in AUR step

+++ src/airootfs/usr/local/lib/arch-grit/modules/99-unmountFilesystems.sh
@@ -3,13 +3,13 @@
 #########################################
 unmountFilesystems() {
     # Remove installation files from new system
-    rm -rf "/mnt/${libPath}" || return 1
+    rm -rf "/mnt/${libPath:?}" || return 1
 
     # Unmount all filesystems to complete the installation
     timeout=0
     until umount -R /mnt ; do
         sleep 5
-        timeout=$(($timeout+5))
+        timeout=$((timeout+5))
         [[ $timeout = 60 ]] && exit 1
     done
 }
